ASINT-2163: Refactor exception in jenkins
