+++ ptai-data-structures/src/main/java/com/ptsecurity/appsec/ai/ee/scan/errors/SettingsMustBeSetUpException.java
@@ -1,4 +1,4 @@
-package com.ptsecurity.appsec.ai.ee.utils.ci.integration.plugin.jenkins.exceptions;
+package com.ptsecurity.appsec.ai.ee.scan.errors;
 
 public class SettingsMustBeSetUpException extends RuntimeException {
     public SettingsMustBeSetUpException(String message) {
